\name{likelihood-package}
\alias{likelihood-package}
\alias{likelihood}
\docType{package}
\title{
Package for maximum likelihood estimation
}
\description{
This package allows you to find the maximum likelihood estimates of statistical models using simulated annealing, a global optimization algorithm.
}
\details{
\tabular{ll}{
Package: \tab likelihood\cr
Version: \tab 1.5\cr
Date: \tab 2012-01-27\cr
Depends: \tab R (>= 2.1.1, nlme\cr
License: \tab GNU Public License\cr
}

Several demonstration scripts are included in the demo directory.
}
\author{
Lora Murphy <murphyl@caryinstitute.org>
Maintainer: Lora Murphy <murphyl@caryinstitute.org>
}
